feat: Split Rake into binary and library
