+++ src/constants.rs
@@ -1,0 +1,2 @@
+pub const MAX_ITER_COUNT: u64 = 0xffffffffffff; // TODO: Temporarily until we handle the math better.
+pub const PREPOSITIONS: [&'static str; 8] = ["as", "of", "for", "by", "like", "in", "from", "into"];
